northbound-seo/backend/pyproject.toml
[tool.poetry]
name = "northbound-seo-backend"
version = "0.1.0"
description = "Northbound SEO Intelligence Platform API"
authors = ["Northbound Growth Co"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.32"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.30"
alembic = "^1.14"
pydantic = "^2.10"
pydantic-settings = "^2.7"
httpx = "^0.28"
anthropic = "^0.40"
firebase-admin = "^6.6"
playwright = "^1.49"
weasyprint = "^63"
python-multipart = "^0.0.20"
beautifulsoup4 = "^4.12"
lxml = "^5.3"
python-slugify = "^8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8"
pytest-asyncio = "^0.24"
httpx = "^0.28"
ruff = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
